feat: update HomePage tab index and add keys for form fields; implement tests for ItemsService and WeatherService

+++ lib/widgets/weather_display.dart
@@ -196,7 +196,7 @@ class WeatherData {
       temperatureCelsius: json['temperature'].toDouble(),
       description: json['description'],
       humidity: json['humidity'],
-      windSpeed: json['windSpeed'].toDouble(),
+      windSpeed: json['windSpeed'],
       icon: json['icon'],
     );
   }
